\name{pacotestRvineSingleCopula}
\alias{pacotestRvineSingleCopula}

\title{
Testing for (j-1)-th order partial copulas in R-vine copulas
}

\description{
The function can be used to test a single copula in a R-vine copula to be a (j-1)-th order partial copula. To apply the function one needs to provide the data and a specified/estimated R-vine copula model in form of a \link[VineCopula]{RVineMatrix} from the package \link[VineCopula]{VineCopula-package}. Additionally, a pacotest options list which can be generated with the \code{\link{pacotestset}} function needs to be provided.
}

\usage{
pacotestRvineSingleCopula(data, RVM, pacotestOptions, tree, copulaNumber)
}

\arguments{
  \item{data}{
  A (n x d) matrix (or data frame) of [0,1] data (i.e. uniform marigns).
  }
  \item{RVM}{
  An \link[VineCopula]{RVineMatrix} object (\link[VineCopula]{VineCopula-package}) which includes the structure, the pair-copula families and parameters.
  }
  \item{pacotestOptions}{
  A options list generated by the \code{\link{pacotestset}} function or the test type as string, i.e., ECORR or VI.
  }
  \item{tree}{
  The tree number (j>=2) of the copula which should be tested to be a (j-1)-th order partial copula.
  }
  \item{copulaNumber}{
  The number (1<= copulaNumber <= j-1) of the copula in the normalized \link[VineCopula]{RVineMatrix} which should be tested to be a (j-1)-th order partial copula.
  }
}

\value{
A list which can, depending on the chosen test, consist of the following elements:
\item{pVal}{
The p-value of the test.
}
\item{TestStat}{
The value of the test statistic.
}
\item{DecisionTree}{
The decision used to partition the support Omega0 of the conditioning variable W. It is provided as list conisisting of three nodes (\code{CentralNode}, \code{LeftNode} and \code{RightNode}) and the variable \code{LeavesForFinalComparison}. Each node consists of the \code{Variable} used to perform the split, the corresponding \code{Quantile} and \code{Threshold}.
}
\item{S}{
The boostrapped values of the test statistic (only for the test type \code{VI}).
}
}

\author{
Malte Kurz
}

\seealso{
\code{\link{pacotest-package}}, \code{\link{pacotest}}, \code{\link{pacotestset}}, \code{\link{pacotestRvineSeq}}
}

\references{
Kurz, M. S. and F. Spanhel (2016), "Testing the simplifying assumption in high-dimensional vine copulas", Unpublished Working Paper.
}

\examples{
# Sample data and R-vine copula selection taken
# from the documentation of RVineStructureSelect
# of the VineCopula package.

data(daxreturns, package ="VineCopula")
dataSet = daxreturns[1:750,1:4]
RVM = VineCopula::RVineStructureSelect(dataSet, c(1:6), progress = TRUE)

# Specify a pacotestOptions list:
# For illustrating the functioning of the decision tree,
# grouped scatterplots and a decision tree plot are activated.
pacotestOptions = pacotestset(testType='ECORR',
                               groupedScatterplots = TRUE,
                               decisionTreePlot = TRUE)

# Test for a 2-nd order partial copula
# corresponding to the variables BMW.DE,BMW.DE
# and conditioning set ALV.DE,BAS.DE
tree = 3
copulaNumber = 1

pacotestResultList = pacotestRvineSingleCopula(dataSet, RVM,
                                                pacotestOptions, tree, copulaNumber)
}

