\name{SAtest}
\alias{SAtest}

\title{Testing the simplifying assumption for a single bivariate conditional copula}

\description{ The function can be used to test the simplifying assumption for a bivariate conditional copula using different tests. Three different test types, the Equal Copula (EC) test, the Equal Rank Correlation (ERC) test and the Vectorial Independence (VI) test are implemeted. For all tests different options can be set by generating a SAtestOptions struct using the \code{\link{SAtestSet}} function.}

\details{
## Applying a test with default options (cf. SAtestSet).
           out = SAtest(U,W,'EC')
           out = SAtest(U,W,'ERC')
           out = SAtest(U,W,'VI')

## Applying a test with options specified in SAtestOptions
           out = SAtest(U,W,SAtestOptions)
}

\arguments{ 
\item{U}{ A (n x 2) matrix of [0,1] data (probability integral transforms), which are the arguments of the  conditional copula of (Y,Z)|W for which the simplifying assumption should be tested. The first column is given by the conditional distribution function of Y|W evaluated at the observed values of Y and W. Analogously, the second column is defined as the conditional distribution function of Z|W evaluated at the observed values of Z and W.}

\item{W}{ A (n x m) vector of observed values for the vector of random variables on which the conditioning is done.
}

\item{SAtestOptions}{ A options struct generated by the \code{\link{SAtestSet}} function or the test type as string, i.e., EqualCop | EC | EqualRankCorr | ERC | VecIndep | VI.}
}

\value{ A list which can, depending on the chosen test, consist of the following elements:
\item{pVal}{ The p-value of the test.}

\item{TestStat}{ The value of the test statistic.}

\item{S}{ The boostrapped values of the test statistic.}

\item{pVals}{ The p-values which are obtained by splitting the observations several time randomly into a training and evaluation data set. The training data set is used to find a sample splitting using decision trees and the test is applied to the evaluation data. The observed p-values are aggregated using the approach of van de Wiel et al. (2009) and Meinhausen et al. (2009).}
}

\author{
Malte Kurz
}

\references{
Hobaek-Haff, I., K. Aas and A. Frigessi (2010), "On the simplified pair-copula construction -- Simply useful or too simplistic?", Journal of Multivariate Analysis 101(5), pp. 1296-1310.

Kojadinovic, I. and M. Holmes (2009), "Tests of independence among continuous random vectors based on Cramer-von Mises functionals of the empirical copula process", Journal of Multivariate Analysis 100(6), pp. 1137-1154.

Kurz, M. and F. Spanhel (2015), "Testing the simplifying assumption in vine copulas", Unpublished Working Paper.

Meinshausen, N., L. Meier, and P. Buehlmann (2009), "p-Values for High-Dimensional Regression", Journal of the American Statistical Association 104(488), pp. 1671-1681.

Omelka, M. and M. Pauly (2012), "Testing equality of correlation coefficients in two populations via permutation methods, Journal of Statistical Planning and Inference 142, pp. 1396-1406.

Quessy, J.-F. (2010), "Applications and asymptotic power of marginal-free tests of stochastic vectorial independence", Journal of Statistical Planning and Inference 140(11), pp. 3058-3075.

Remillard, B. and O. Scaillet (2009), "Testing for equality between two copulas", Journal of Multivariate Analysis 100(3), pp. 377-386.

Spanhel, F. and M. Kurz (2015), "Simplified vine copula models: Approximations based on the simplifying assumption", Unpublished Working Paper.

van de Wiel, M. A., J. Berkhof, and W. N. van Wieringen (2009), "Testing the prediction error difference between 2 predictors", Biostatistics 10(3), pp. 550-560.
}

\seealso{\code{\link{SAtest-package}}, \code{\link{SAtestSet}}
}


