\name{pacotestset}
\alias{pacotestset}

\title{Create / alter pacotest options list}

\description{ The function creates or updates a list object, which is required for applying the \code{\link{pacotest}} function.}

\details{
Calling without any arguments prints all possible options.
           \preformatted{pacotestset}

Calling with a string, that specifies the test type, gives back a option list with the default values corresponding to each test.
           \preformatted{pacotestOptions = pacotestset('EC')}
           \preformatted{pacotestOptions = pacotestset('ERC')}
           \preformatted{pacotestOptions = pacotestset('VI')}

Calling with pairs of parameter names and values creates an pacotestOptions list in which the named parameters have the specified values.
           \preformatted{pacotestOptions = pacotestset('Name1',Value1,'Name2',Value2,...)}

Calling with an existing pacotestOptions list checks the list for consistency.
           \preformatted{pacotestset(pacotestOptions);}

Calling with an existing pacotestOptions structuer and pairs of parameter names and values creates a copy of the existing list, where the named parameters are updated with the provided values.
           \preformatted{pacotestOptions = pacotestset(pacotestOptions,'Name1',Value1,'Name2',Value2,...)}
}

\arguments{ 
\item{pacotestOptions}{A options list for the \code{\link{pacotest}} function generated by the \code{\link{pacotestset}} function.}

\item{TestType}{A string which specifies the type of the test for testing the simplifying assumption (SA).

Possible values: EqualCop | EC | EqualRankCorr | ERC | VecIndep | VI
}

\item{Grouping}{For \code{TestType = EqualCop | EC} and \code{TestType = EqualRankCorr | ERC}:

The grouping method which is used to form two heterogeneous groups using the information contained in the conditioning variable of the conditional copula.

Possible values: TreeERC | TreeEC | SumMedian | SumThirdsI | SumThirdsII | ProdMedian | ProdThirdsI | ProdThirdsII
}

\item{AggPvalsNumbRep}{For \code{TestType = EqualRankCorr | ERC}:

The number of replications for the van de Wiel et al. (2009) and Meinhausen et al. (2009) approach.
}

\item{GroupedScatterplots}{For \code{TestType = EqualCop | EC}:

A logical whether grouped scatterplots should be produced.

For \code{TestType = EqualRankCorr | ERC}:

A logical whether grouped scatterplots should be produced. This option can not be enabled if AggPvalsNumbRep is larger than one.
}

\item{ExpMinSampleSize}{For \code{TestType = EqualCop | EC} and \code{TestType = EqualRankCorr | ERC}:

The minimum number of observations which are allocated to a group in the decision tree learning process. Under the i.i.d. assumption for the data generating process it is also the minimum expected number of observations allocated to a specific group in the evaluation data set. The default value is 50.
}

\item{TrainingDataFraction}{For \code{TestType = EqualCop | EC} and \code{TestType = EqualRankCorr | ERC}:

The fraction of data which is used to train the decision tree. The default value is 0.5.
}


\item{NumbBoot}{For \code{TestType = EqualCop | EC} and \code{TestType = VecIndep | VI}:

The number of bootstrap replications for computing p-values using the multiplier bootstrap approach.
}

}

\value{The function returns a \code{pacotestOptions} list which can be used as input argument of the function \code{\link{pacotest}}
}
\author{
Malte Kurz
}
\references{
Hobaek-Haff, I., K. Aas and A. Frigessi (2010), "On the simplified pair-copula construction -- Simply useful or too simplistic?", Journal of Multivariate Analysis 101(5), pp. 1296-1310.

Kojadinovic, I. and M. Holmes (2009), "Tests of independence among continuous random vectors based on Cramer-von Mises functionals of the empirical copula process", Journal of Multivariate Analysis 100(6), pp. 1137-1154.

Kurz, M. and F. Spanhel (2015), "Testing the simplifying assumption in vine copulas", Unpublished Working Paper.

Meinshausen, N., L. Meier, and P. Buehlmann (2009), "p-Values for High-Dimensional Regression", Journal of the American Statistical Association 104(488), pp. 1671-1681.

Omelka, M. and M. Pauly (2012), "Testing equality of correlation coefficients in two populations via permutation methods, Journal of Statistical Planning and Inference 142, pp. 1396-1406.

Quessy, J.-F. (2010), "Applications and asymptotic power of marginal-free tests of stochastic vectorial independence", Journal of Statistical Planning and Inference 140(11), pp. 3058-3075.

Remillard, B. and O. Scaillet (2009), "Testing for equality between two copulas", Journal of Multivariate Analysis 100(3), pp. 377-386.

Spanhel, F. and M. Kurz (2015), "Simplified vine copula models: Approximations based on the simplifying assumption", Unpublished Working Paper.

van de Wiel, M. A., J. Berkhof, and W. N. van Wieringen (2009), "Testing the prediction error difference between 2 predictors", Biostatistics 10(3), pp. 550-560.
}

\seealso{\code{\link{pacotest-package}}, \code{\link{pacotest}}
}


